+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0]
+
+### Added
+
+- Option to add attached files to Digital Shipments.
+- API endpoint to get attached files from a shipment.
+
+### Changed
+
+- Updated nginx configuration.
+- Made Memo Label optional for Fjernpost.
+- Made shipment order of letters stable to ensure proper caching.
+
 ## [0.2.0]
 
 ### Added
@@ -41,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.3.0
 [0.2.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.2.0
 [0.1.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.1.0
 [0.0.1]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.0.1